pyproject.toml
[project]
name = "escalada-api"
version = "0.1.0"
description = "Escalada API (FastAPI/WebSocket) + persistence layer"
authors = [
    {name = "Silviu"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "reportlab (>=4.4.0,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.30,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)",
    "alembic (>=1.13.1,<2.0.0)",
    "pydantic-settings (>=2.2.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
isort = "^7.0.0"
pre-commit = "^4.5.1"
pytest-asyncio = "^1.3.0"
